fix: prevent repeated form transmissions
We’ve attached form submit and submit button click event handlers but we
only need the former, as the click event triggers the submit event
anyway. Choosing the submit event is the right thing here, because it
does not rely on the button being clicked and therefore works just as
well with keyboard interactions.