feat: Inform capsule owner of inactive recipients
Running management command "sendcapsulehints" results in an email sent
to capsule owners with inactive receivers. "Inactivity" means that the
receiver is at least settings.INACTIVE_RECEIVER_HINT_DAYS old and
matches the criteria defined by CapsuleReceiver.is_active().

The command sends the emails on every run. We want to avoid having
a field for each kind of reminder on the model. Thus the frequency of
running the command determines the frequency of emails sent.

Refs #26